Amend date of manual workflow run from 2021-01-04 to 2021-01-03
Issue was caused by breaking change to ad-m/github-push-action
that pushed data for 2021-01-02 to main rather than master.

Data from 2021-01-02 was merged in via PR #2. Data for 2021-01-03
was failing to be updated automatically. Thus, a manual run that
doesn't rely on ad-m/github-push-action was run and dates are
amended accordingly.